+++ Package.swift
@@ -15,8 +15,8 @@ let package = Package(
     ],
     targets: [
         .binaryTarget(name: "FireworkVideo",
-                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v0.4.0/FireworkVideo-v0.4.0.xcframework.zip",
-                      checksum: "cfd7894407e6f76598e7edcbd22694dcc79a5b2c3a3edd606e5e65ced29d52f8")
+                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v0.5.0/FireworkVideo-v0.5.0.xcframework.zip",
+                      checksum: "0b8ed086711906f20b5ef594774d7b984e2cebd1969f2ba73e4aa71b8dfe7b2d"),
     ]
     
 )