+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2021-04-01
+
+### Added
+
+- Configurable sharing functionality in video player.
+- Playback delegate event for share button tap.
+- Feed delegate for video thumbnail tap.
+
+### Changed
+
+- Renamed `VideoPlayerContentConfiguration.FadeOutDelay.none` to `VideoPlayerContentConfiguration.FadeOutDelay.disabled` to remove ambiguity.
+- Better support for launching the video player from landscape orientation. 
+
 ## [0.4.0] - 2021-03-18
 
 ### Added